controller_node/requirements.txt
pyserial
pyzmq
numpy
opencv-python